requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-jazzmin==2.6.0
python-dotenv==1.1.0
dj-database-url==3.0.0 